+++ Sources/Plasma/PubUtilLib/plStatGather/plCalculatedProfiles.cpp
@@ -65,8 +65,8 @@ plProfile_CreateCounter("Polys Per Material", "General", PolysPerMat);
 void CalculateProfiles()
 {
     // KLUDGE - do timing that overlaps the beginframe / endframe (where timing is normally reset)
-    static uint32_t lastTicks = plProfileManager::GetTime();
-    uint32_t curTicks = plProfileManager::GetTime();
+    static uint64_t lastTicks = plProfileManager::GetTime();
+    uint64_t curTicks = plProfileManager::GetTime();
     gVarRFPS.Set(curTicks - lastTicks);
     lastTicks = curTicks;
 
@@ -129,9 +129,9 @@ void UpdateStandardGraphs(float xPos, float yPos)
     if (fFPSPlate)
     {
         fFPSPlate->AddData(
-            gVarRFPS.GetValue(),
-            plProfile_GetValue(DrawTime),
-            plProfile_GetValue(UpdateTime));
+            (int32_t)gVarRFPS.GetValue(),
+            (int32_t)plProfile_GetValue(DrawTime),
+            (int32_t)plProfile_GetValue(UpdateTime));
         PositionPlate(fFPSPlate);
     }
 }
